Refactored models out of db_access_v2
