+++ functions/pipeline/shared/db_access/__init__.py
@@ -1,1 +1,2 @@
-from .db_access_v2 import ImageTagDataAccess, ImageTag, ImageInfo, ImageTagState, ImageLabel, PredictionLabel
+from .models import ImageTag, ImageInfo, ImageTagState, ImageLabel, PredictionLabel
+from .db_access_v2 import ImageTagDataAccess
